renamed CreatedTripsScreen to MyCreatedTripsScreen.
