+++ src/screens/MyCreatedTripsScreen.js
