+++ src/screens/index.js
@@ -8,5 +8,5 @@ export {default as HomeScreen} from './HomeScreen'
 export {default as CreateTripScreen} from './CreateTripScreen'
 export {default as MyProfileScreen} from './MyProfileScreen'
 export {default as TripsScreen} from './TripsScreen'
-export {default as CreatedTripsScreen} from './CreatedTripsScreen'
+export {default as MyCreatedTripsScreen} from './MyCreatedTripsScreen'
 export {default as EditTripScreen} from './EditTripScreen'
